ICPC题目提取schema/requirements.txt
pdfplumber>=0.11.5
pydantic>=2.7.0
requests>=2.32.0
python-dotenv>=1.0.1
